Fix incorrect returned values from pin_degreesFromOrientation
Whilst the pin_degreesFromOrientation function was the correct amount a pixel grid had been rotated. How it was used was the opposite, instead of undoing the baked pixel grid rotation. It applied it again. So, images that were LEFT or RIGHT (or mirrored of the two) would appear upside down.